Moved local json to dedicated folder
